+ Changing type for gameTimeMS

+++ kex2/turok/game/gameManager.h
@@ -115,7 +115,7 @@ private:
     wipecallback_t          *wipeCallback;
     void                    *wipeCallbackData;
 
-    int                     gameTimeMS;
+    uint64_t                gameTimeMS;
 
     bool                    bPaused;
 };
